+++ tests/rules/sentence-case-heading.test.js
@@ -7,25 +7,28 @@ import fs from 'fs';
 import path from 'path';
 import { fileURLToPath } from 'url';
 import { describe, test, expect } from '@jest/globals';
-import debug from '../logger.js';
+import debug from '../../logger.js';
 
 // Import markdownlint using the proper ES modules path
 import { lint } from 'markdownlint/promise';
 
 // Import the custom rule
-import sentenceCaseHeadingRule from '../.vscode/custom-rules/sentence-case-heading.js';
-import { parseFixture } from './utils/fixture.js';
+import sentenceCaseHeadingRule from '../../.vscode/custom-rules/sentence-case-heading.js';
+import { parseFixture } from '../utils/fixture.js';
 
 // Get current file path (ES modules don't have __dirname)
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
 
 // Path to the fixture file
-const fixturePath = path.join(__dirname, "sentence-case-edge-cases.fixture.md");
+const fixturePath = path.join(
+  __dirname,
+  '../fixtures/basic-sentence-case-heading.fixture.md'
+);
 
 
 
-describe("sentence-case-heading edge cases", () => {
+describe("sentence-case-heading rule", () => {
   const { passingLines, failingLines } = parseFixture(fixturePath);
   const log = debug.extend('tests');
   const fixtureLines = fs.readFileSync(fixturePath, 'utf8').split('\n');
@@ -69,12 +72,14 @@ describe("sentence-case-heading edge cases", () => {
     })));
   });
 
-  headingCases.forEach(({ lineNumber, headingText, expectViolation }) => {
-    test(`line ${lineNumber}: "${headingText}"`, () => {
+  test.each(headingCases)(
+    'line %i: "%s"',
+    ({ lineNumber, headingText, expectViolation }) => {
       const hasViolation = ruleViolations.some(v => v.lineNumber === lineNumber);
       expect(hasViolation).toBe(expectViolation);
-    });
-  });
+    }
+  );
+  
   test("provides appropriate error messages", async () => {
     const options = {
       customRules: [sentenceCaseHeadingRule],
@@ -110,5 +115,9 @@ describe("sentence-case-heading edge cases", () => {
     });
 
     const fixtureLines = fs.readFileSync(fixturePath, "utf8").split("\n");
+    const cssLine = fixtureLines.findIndex(line => line.startsWith("# css")) + 1;
+    const cssViolation = ruleViolations.find(v => v.lineNumber === cssLine);
+    expect(cssViolation).toBeTruthy();
+    expect(cssViolation.errorDetail).toBe("Single-word heading should be capitalized.");
   });
 });
